refactor: refactor new header into components
